+++ app/assets/javascripts/turbo-mount/svelte.js
@@ -1,26 +1,15 @@
-import { TurboMountController, TurboMount } from 'turbo-mount';
+import { buildRegisterFunction } from 'turbo-mount';
+export { TurboMount } from 'turbo-mount';
 
-class TurboMountSvelteController extends TurboMountController {
-    constructor() {
-        super(...arguments);
-        this.framework = "svelte";
-    }
-    mountComponent(el, Component, props) {
+const plugin = {
+    mountComponent: (mountProps) => {
+        const { el, Component, props } = mountProps;
         const component = new Component({ target: el, props });
         return () => {
             component.$destroy();
         };
-    }
-}
-
-const plugin = {
-    framework: "svelte",
-    controller: TurboMountSvelteController,
+    },
 };
-class TurboMountSvelte extends TurboMount {
-    constructor(props) {
-        super(Object.assign(Object.assign({}, props), { plugin }));
-    }
-}
+const registerComponent = buildRegisterFunction(plugin);
 
-export { TurboMountSvelte as TurboMount, TurboMountSvelte, plugin as default };
+export { plugin as default, registerComponent };