+++ packages/turbo-mount/src/plugins/svelte/registerComponents.ts
@@ -1,0 +1,6 @@
+import { buildRegisterComponentsFunction } from "turbo-mount/registerComponents";
+import plugin from "turbo-mount/svelte";
+
+const registerComponents = buildRegisterComponentsFunction(plugin);
+
+export { registerComponents };