+++ packages/turbo-mount/src/plugins/vue/registerComponents.ts
@@ -1,0 +1,6 @@
+import { buildRegisterComponentsFunction } from "turbo-mount/registerComponents";
+import plugin from "turbo-mount/vue";
+
+const registerComponents = buildRegisterComponentsFunction(plugin);
+
+export { registerComponents };